+++ anvil_consortium_manager/tests/test_app/adapters.py
@@ -1,7 +1,7 @@
 from anvil_consortium_manager.adapters.account import BaseAccountAdapter
 from anvil_consortium_manager.adapters.workspace import BaseWorkspaceAdapter
 
-from . import forms, models, tables
+from . import filters, forms, models, tables
 
 
 class TestWorkspaceAdapter(BaseWorkspaceAdapter):
@@ -11,6 +11,7 @@ class TestWorkspaceAdapter(BaseWorkspaceAdapter):
     type = "test"
     description = "Workspace type for testing"
     list_table_class = tables.TestWorkspaceDataTable
+    workspace_form_class = forms.TestWorkspaceForm
     workspace_data_model = models.TestWorkspaceData
     workspace_data_form_class = forms.TestWorkspaceDataForm
     workspace_detail_template_name = "test_workspace_detail.html"
@@ -29,6 +30,7 @@ class TestAccountAdapter(BaseAccountAdapter):
     """Test adapter for accounts."""
 
     list_table_class = tables.TestAccountTable
+    list_filterset_class = filters.TestAccountListFilter
 
     def get_autocomplete_queryset(self, queryset, q):
         if q: