+++ example_site/app/models.py
@@ -4,7 +4,7 @@ from anvil_consortium_manager.models import BaseWorkspaceData
 
 
 # Create your models here.
-class ExampleWorkspaceData(BaseWorkspaceData):
+class CustomWorkspaceData(BaseWorkspaceData):
     """Example custom model to hold additional data about a Workspace."""
 
     study_name = models.CharField(max_length=255)